requirements.txt
# Tech Challenge Fase 4 - Dependências
# Python 3.12+

# Core - Visão Computacional
opencv-python>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0

# Análise de Emoções (FER)
fer>=22.5.0

# Detecção de Atividades (YOLO11-pose)
ultralytics>=8.3.0

# Interface Gráfica
customtkinter>=5.2.0
matplotlib>=3.8.0

# Utilitários
python-dotenv>=1.0.0

